Discourage users from disabling local caching
It is of high importance that clients use local cache whenever and wherever possible. This has direct impact on the overall load towards central services. Therefore, users shall be encouraged to use local caching and do not disable it unless it is strictly required.

In addition, this change proposal marks the other optional action (iasToXsuaaTokenExchange) as such, to keep the same approach on bullet and section naming.